+++ package.json
@@ -62,6 +62,7 @@
     "connectivity": "yarn workspace @sap-cloud-sdk/connectivity",
     "generator": "yarn workspace @sap-cloud-sdk/generator",
     "generator-common": "yarn workspace @sap-cloud-sdk/generator-common",
+    "http-client": "yarn workspace @sap-cloud-sdk/http-client",
     "openapi-generator": "yarn workspace @sap-cloud-sdk/openapi-generator",
     "odata-common": "yarn workspace @sap-cloud-sdk/odata-common",
     "odata-v2": "yarn workspace @sap-cloud-sdk/odata-v2",
